Refactor: api 관심사별로 분류 및 파일경로 변경
