+++ components/best-articles.tsx
@@ -1,11 +1,11 @@
 import { useEffect, useState } from 'react';
-import { getBestArticles } from '@/lib/api/getArticles';
+import { getBestArticles } from '@/lib/api/articles';
 import { ArticleProps } from '@/types';
 import getFormatDate from '@/lib/utils/formatDate';
 import Link from 'next/link';
 import Image from 'next/image';
 import imageBadge from '@/public/images/img_badge.svg';
-import icHeart from '@/public/images/icons/ic_heart.png';
+import icEmptyHeart from '@/public/images/icons/ic_empty-heart.svg';
 import { constants, IS_SERVER } from '../lib/constants';
 
 function getPageSize() {
@@ -75,7 +75,7 @@ function BestArticlePreview({ createdAt, likeCount, image, title, writer }: Arti
         <div className='flex items-center gap-2'>
           <span className='text-cool-gray600'>{writer.nickname}</span>
           <div className='flex items-center gap-1'>
-            <Image src={icHeart} alt='좋아요 하트' width={16} height={16}></Image>
+            <Image src={icEmptyHeart} alt='좋아요 하트' width={16} height={16}></Image>
             <span className='text-cool-gray500'>{likeCount}</span>
           </div>
         </div>
