+++ lib/api/articles.ts
@@ -1,4 +1,4 @@
-import { ArticleProps, ArticleList } from '@/types';
+import { ArticleProps, ArticleList, PostArticleType } from '@/types';
 import { axiosInstance } from './axios';
 
 export const getArticles = async (
@@ -25,5 +25,28 @@ export const getBestArticles = async (page: number, pageSize: number, orderBy: s
   } catch (error) {
     console.error(`Failed to fetch items: ${error}`);
     throw new Error('정보를 불러오는데 실패했습니다');
+  }
+};
+
+export const postArticle = async (
+  title: string,
+  content: string,
+  token: string,
+  image?: string | undefined
+): Promise<PostArticleType> => {
+  try {
+    const response = await axiosInstance.post(
+      `/articles`,
+      { content, title, image },
+      {
+        headers: {
+          Authorization: `Bearer ${token}`,
+        },
+      }
+    );
+    return response.data;
+  } catch (error) {
+    console.error(`Failed to post comment: ${error}`);
+    throw new Error('댓글 작성에 실패했습니다');
   }
 };
