+++ lib/api/axios.ts
@@ -1,5 +1,4 @@
 import axios from 'axios';
-import { postRefreshToken } from './postRefreshToken';
 
 export const baseURL = process.env.NEXT_PUBLIC_BASE_URL;
 
@@ -9,5 +8,3 @@ export const axiosInstance = axios.create({
     'Content-Type': 'application/json',
   },
 });
-
-postRefreshToken(axiosInstance);
